requirements.txt
asgiref==3.8.1
boto3==1.35.68
botocore==1.35.68
dj-database-url==2.3.0
Django==5.0.6
django-cors-headers==4.4.0
django-heroku==0.3.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
